+++ Konsole.Sample/Program.cs
@@ -17,15 +17,6 @@ namespace Konsole.Sample
     class Program
     {
         
-        // random notes and links
-
-        // experiment to see if we can read a block of the screen by using move and redirecting input and output
-        // see http://stackoverflow.com/questions/12355378/read-from-location-on-console-c-sharp
-
-        // windows specific 
-        // more information https://msdn.microsoft.com/en-us/library/windows/desktop/ms681913(v=vs.85).aspx
-
-
         private static void RandomStuff(IConsole con)
         {
             var pb = new ProgressBar(con, PbStyle.DoubleLine, 50);
@@ -33,45 +24,6 @@ namespace Konsole.Sample
             Console.ReadKey(true);
             pb.Max = 40;
             Console.ReadKey(true);
-        }
-
-
-        private static void RunCommand(IConsole con, char cmd)
-        {
-            switch (cmd)
-            {
-                case 'f':
-                    FormDemos.Run(con);
-                    break;
-
-                case 'w':
-                    WindowDemo.Run(con);
-                    break;
-
-                case 'r':
-                    RandomStuff(con);
-                    break;
-
-                case 'h':
-                    HiliteDemo.Run(con);
-                    break;
-
-
-                case '2':
-                    ProgressBarDemos.ParallelConstructorDemo();
-                    break;
-
-
-                case 'p':
-                    ProgressBarDemos.SimpleDemo(con);
-                    break;
-
-
-
-                default:
-                    break;
-            }
-
         }
 
         private static void Main(string[] args)
@@ -85,8 +37,8 @@ namespace Konsole.Sample
                 new MenuItem('f', "Forms", ()=> FormDemos.Run(output1)),
                 new MenuItem('b', "Boxes", ()=> BoxeDemos.Run(output1)),
                 new MenuItem('s', "Scrolling", ()=> WindowDemo.Run2(con)),
-                new MenuItem('p', "ProgressBar1", ()=> ProgressBarDemos.ParallelDemo(con)),
-                new MenuItem('q', "ProgressBar2", () => ProgressBarDemos.ParallelDemo(con)),
+                new MenuItem('p', "ProgressBar 1 line demo", ()=> ProgressBarDemos.ParallelDemo(con)),
+                new MenuItem('q', "ProgressBar 2 line demo", () => ProgressBarDemos.ParallelDemo(con)),
                 new MenuItem('t', "Test data", () => TestDataDemo.Run(con)),
                 new MenuItem('c', "clear screen", () => con.Clear()),
                 new MenuItem('r', "RANDOM", () => RandomStuff(con))