+++ Konsole/IConsole.cs
@@ -1,9 +1,10 @@
 using System;
+using Konsole.Drawing;
 using Konsole.Internal;
+using Konsole.Menus;
 
 namespace Konsole
 {
-
     public interface IConsole : IWriteLine
     {
         ConsoleState State { get; set; }
@@ -22,11 +23,7 @@ namespace Konsole
         void PrintAtColor(ConsoleColor foreground, int x, int y, string text, ConsoleColor? background);
         void ScrollUp();
         void Clear();
-        IConsole TopHalf(ConsoleColor foreground = ConsoleColor.Cyan, ConsoleColor background = ConsoleColor.Black);
-        IConsole BottomHalf(ConsoleColor foreground = ConsoleColor.Green,ConsoleColor background = ConsoleColor.Black);
 
-        // saves the current cursor position, (left and top) as well as colors
-        // runs the command, then restores the state.
         void MoveBufferArea(int sourceLeft, int sourceTop, int sourceWidth, int sourceHeight, int targetLeft,
             int targetTop, char sourceChar, ConsoleColor sourceForeColor, ConsoleColor sourceBackColor);
     }