+++ Konsole/Properties/AssemblyInfo.cs
@@ -34,3 +34,5 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("2.0.2.*")]
 
+// this here so that we can do some simple testing of some of the internals while spiking new fixes.
+[assembly: InternalsVisibleTo("TestPackage.V3.0.0")]